pyproject.toml
[tool.poetry]
name = "project-mobile-tests"
version = "0.1.0"
description = ""
authors = ["viktoriya <tkvika@yandex.ru>"]
readme = "README.md"
packages = [{include = "project_mobile_tests"}]

[tool.poetry.dependencies]
python = "^3.10"
selene = {version = "^2.0.0-beta.8", allow-prereleases = true}
Appium-Python-Client = "^2.6.1"
pytest = "^7.1.3"
allure-pytest = "^2.10.0"
python-dotenv = "^0.21.0"

[tool.pytest.ini_options]
addopts = "--alluredir allure-results --clean-alluredir "

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
